Flatten module directories into single files for simpler structure
Consolidate multi-file module directories into single .rs files:
- cell/{mod.rs,io.rs} -> cell.rs
- library/mod.rs -> library.rs
- point/mod.rs -> point.rs
- grid/mod.rs -> grid.rs
- traits/mod.rs -> traits.rs
- elements/polygon/{mod.rs,io.rs,utils.rs} -> elements/polygon.rs
- elements/path/{mod.rs,io.rs,path_type/mod.rs} -> elements/path.rs
- elements/reference/{mod.rs,io.rs,instance/mod.rs} -> elements/reference.rs
- elements/text/{mod.rs,io.rs,utils.rs,presentation/mod.rs} -> elements/text.rs

Updated import paths in lib.rs and utils/io.rs accordingly.
