+++ crates/gdsr/src/elements/reference.rs
@@ -1,11 +1,78 @@
+use std::sync::Arc;
 use std::sync::mpsc;
 
-use crate::{Element, Grid, Library, Movable, Point, Transformable, Transformation};
+use crate::config::gds_file_types::{GDSDataType, GDSRecord, combine_record_and_data_type};
+use crate::elements::Element;
+use crate::error::GdsError;
+use crate::traits::{Movable, ToGds, Transformable};
+use crate::utils::io::{
+    validate_col_row, write_element_tail_to_file, write_points_to_file,
+    write_string_with_record_to_file, write_transformation_to_file, write_u16_array_to_file,
+};
+use crate::{Cell, Grid, Library, Point, Transformation};
 
-pub mod instance;
-mod io;
+/// The target of a [`Reference`](crate::Reference): either a cell name or an inline element.
+#[derive(Clone, Debug, PartialEq)]
+pub enum Instance {
+    /// A reference to a named cell in the library.
+    Cell(String),
+    /// An inline element stored directly in the reference.
+    Element(Arc<Box<Element>>),
+}
 
-pub use instance::Instance;
+impl Instance {
+    /// Returns the cell name if this is a `Cell` instance, or `None`.
+    pub fn as_cell(&self) -> Option<&String> {
+        if let Self::Cell(v) = self {
+            Some(v)
+        } else {
+            None
+        }
+    }
+
+    /// Returns the element if this is an `Element` instance, or `None`.
+    pub fn as_element(&self) -> Option<&Arc<Box<Element>>> {
+        if let Self::Element(v) = self {
+            Some(v)
+        } else {
+            None
+        }
+    }
+}
+
+impl Default for Instance {
+    fn default() -> Self {
+        Self::Cell(String::new())
+    }
+}
+
+impl std::fmt::Display for Instance {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::Cell(name) => write!(f, "Cell instance: {name}"),
+            Self::Element(element) => write!(f, "Element instance: {element}"),
+        }
+    }
+}
+
+// From implementations for common element types
+impl From<&Cell> for Instance {
+    fn from(value: &Cell) -> Self {
+        Self::Cell(value.name().to_string())
+    }
+}
+
+impl From<String> for Instance {
+    fn from(value: String) -> Self {
+        Self::Cell(value)
+    }
+}
+
+impl From<&str> for Instance {
+    fn from(value: &str) -> Self {
+        Self::Cell(value.to_string())
+    }
+}
 
 /// A reference to an instance (cell or element) placed with a grid layout.
 #[derive(Clone, Debug, PartialEq, Default)]
@@ -226,10 +293,246 @@ impl Movable for Reference {
     }
 }
 
+impl ToGds for Reference {
+    fn to_gds_impl(&self, database_units: f64) -> Result<Vec<u8>, GdsError> {
+        match &self.instance {
+            Instance::Cell(cell_name) => self.to_gds_impl_with_cell(database_units, cell_name),
+            Instance::Element(element) => {
+                self.to_gds_impl_with_element(database_units, element.as_ref().as_ref())
+            }
+        }
+    }
+}
+
+impl Reference {
+    fn to_gds_impl_with_element(
+        &self,
+        database_units: f64,
+        element: &Element,
+    ) -> Result<Vec<u8>, GdsError> {
+        let grid = self.grid();
+        let spacing_x = grid.spacing_x().unwrap_or_default();
+        let spacing_y = grid.spacing_y().unwrap_or_default();
+
+        let mut buf = Vec::new();
+        for column_index in 0..grid.columns() {
+            for row_index in 0..grid.rows() {
+                let offset = (spacing_x * column_index) + (spacing_y * row_index);
+                let rotated_offset =
+                    offset.rotate_around_point(grid.angle(), &crate::Point::default());
+                let final_position = grid.origin() + rotated_offset;
+
+                let mut new_element = element.clone();
+                if grid.x_reflection() {
+                    new_element = new_element.reflect(0.0, grid.origin());
+                }
+                new_element = new_element.rotate(grid.angle(), grid.origin());
+                new_element = new_element.scale(grid.magnification(), grid.origin());
+                new_element = new_element.move_by(final_position - grid.origin());
+
+                buf.extend_from_slice(&new_element.to_gds_impl(database_units)?);
+            }
+        }
+        Ok(buf)
+    }
+
+    fn to_gds_impl_with_cell(
+        &self,
+        database_units: f64,
+        cell_name: &str,
+    ) -> Result<Vec<u8>, GdsError> {
+        validate_col_row(self.grid().columns(), self.grid().rows())?;
+
+        let is_single_instance = self.grid().columns() == 1 && self.grid().rows() == 1;
+
+        let record = if is_single_instance {
+            GDSRecord::SRef
+        } else {
+            GDSRecord::ARef
+        };
+
+        let mut buffer = Vec::new();
+
+        let buffer_start = [4, combine_record_and_data_type(record, GDSDataType::NoData)];
+
+        write_u16_array_to_file(&mut buffer, &buffer_start)?;
+
+        write_string_with_record_to_file(&mut buffer, GDSRecord::SName, cell_name)?;
+
+        let angle = self.grid().angle();
+        let magnification = self.grid().magnification();
+        let x_reflection = self.grid().x_reflection();
+
+        write_transformation_to_file(&mut buffer, angle, magnification, x_reflection)?;
+
+        if is_single_instance {
+            let origin = self.grid().origin();
+            let reference_points = [origin];
+            write_points_to_file(&mut buffer, &reference_points, database_units)?;
+        } else {
+            let buffer_array = [
+                8,
+                combine_record_and_data_type(GDSRecord::ColRow, GDSDataType::TwoByteSignedInteger),
+                self.grid().columns() as u16,
+                self.grid().rows() as u16,
+            ];
+
+            write_u16_array_to_file(&mut buffer, &buffer_array)?;
+
+            let origin = self
+                .grid()
+                .origin()
+                .rotate_around_point(self.grid().angle(), &self.grid().origin());
+
+            match (self.grid.spacing_x(), self.grid.spacing_y()) {
+                (Some(spacing_x), Some(spacing_y)) => {
+                    let point2 = ((origin + spacing_x) * self.grid().columns())
+                        .rotate_around_point(self.grid().angle(), &origin);
+
+                    let point3 = ((origin + spacing_y) * self.grid().rows())
+                        .rotate_around_point(self.grid().angle(), &origin);
+
+                    let reference_points = [origin, point2, point3];
+                    write_points_to_file(&mut buffer, &reference_points, database_units)?;
+                }
+                (Some(spacing_x), None) => {
+                    let point2 = ((origin + spacing_x) * self.grid().columns())
+                        .rotate_around_point(self.grid().angle(), &origin);
+                    let reference_points = [origin, point2, origin];
+                    write_points_to_file(&mut buffer, &reference_points, database_units)?;
+                }
+                (None, Some(spacing_y)) => {
+                    let point3 = ((origin + spacing_y) * self.grid().rows())
+                        .rotate_around_point(self.grid().angle(), &origin);
+                    let reference_points = [origin, origin, point3];
+                    write_points_to_file(&mut buffer, &reference_points, database_units)?;
+                }
+                _ => {
+                    let reference_points = [origin];
+                    write_points_to_file(&mut buffer, &reference_points, database_units)?;
+                }
+            }
+        }
+
+        write_element_tail_to_file(&mut buffer)?;
+
+        Ok(buffer)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::elements::Polygon;
+
+    mod instance {
+        use super::*;
+        use crate::{Path, Point, Polygon, Text};
+
+        #[test]
+        fn test_instance_cell() {
+            let instance = Instance::from("test_cell");
+            assert_eq!(instance, Instance::Cell("test_cell".to_string()));
+        }
+
+        #[test]
+        fn test_instance_from_string() {
+            let instance = Instance::from(String::from("my_cell"));
+            assert_eq!(instance, Instance::Cell("my_cell".to_string()));
+        }
+
+        #[test]
+        fn test_instance_from_polygon() {
+            let polygon = Polygon::new(
+                [
+                    Point::integer(0, 0, 1e-9),
+                    Point::integer(10, 0, 1e-9),
+                    Point::integer(10, 10, 1e-9),
+                ],
+                1,
+                0,
+            );
+            assert!(Instance::from(polygon).as_element().is_some());
+        }
+
+        #[test]
+        fn test_instance_default() {
+            let instance = Instance::default();
+            assert_eq!(instance, Instance::Cell(String::new()));
+        }
+
+        #[test]
+        fn test_instance_display() {
+            let instance = Instance::from("test_cell");
+            insta::assert_snapshot!(instance.to_string(), @"Cell instance: test_cell");
+        }
+
+        #[test]
+        fn test_instance_clone() {
+            let instance = Instance::from("test_cell");
+            let cloned = instance.clone();
+            assert_eq!(instance, cloned);
+        }
+
+        #[test]
+        fn test_instance_from_path() {
+            let path = Path::new(
+                vec![Point::integer(0, 0, 1e-9), Point::integer(10, 10, 1e-9)],
+                1,
+                0,
+                None,
+                None,
+            );
+
+            let instance = Instance::from(path);
+
+            assert!(instance.as_element().unwrap().as_path().is_some());
+        }
+
+        #[test]
+        fn test_instance_from_text() {
+            use crate::{HorizontalPresentation, VerticalPresentation};
+
+            let text = Text::new(
+                "test",
+                Point::integer(0, 0, 1e-9),
+                1,
+                0,
+                1.0,
+                0.0,
+                false,
+                VerticalPresentation::Middle,
+                HorizontalPresentation::Centre,
+            );
+
+            let instance = Instance::from(text);
+
+            assert!(instance.as_element().unwrap().as_text().is_some());
+        }
+
+        #[test]
+        fn test_instance_from_cell() {
+            let cell = Cell::new("my_cell");
+            let instance = Instance::from(&cell);
+
+            assert_eq!(instance, Instance::Cell("my_cell".to_string()));
+        }
+
+        #[test]
+        fn test_instance_display_element() {
+            let polygon = Polygon::new(
+                [
+                    Point::integer(0, 0, 1e-9),
+                    Point::integer(10, 0, 1e-9),
+                    Point::integer(10, 10, 1e-9),
+                ],
+                1,
+                0,
+            );
+            let instance = Instance::from(polygon);
+            insta::assert_snapshot!(instance.to_string(), @"Element instance: Polygon with 4 point(s), starting at (0 (1.000e-9), 0 (1.000e-9)) on layer 1, data type 0");
+        }
+    }
 
     #[test]
     fn test_reference_new() {
