+++ crates/gdsr/src/grid.rs
