+++ crates/gdsr/src/traits.rs
